Add aarch64 as alias for arm64 (#71)
Add aarch64 to the list of arm64 devices.

Fixes an issue on linux where arm64 is not defined.

Previously, arm64 was only defined as arm64, which only matches modern
macOS on apple silicon.

As it turns out, there other arm64 devices besides apple silicon, and
being able to compile against them is actually kind of nice. This PR
adds detection for aarch64, which is also considered arm64 and allows
V8 to compile and run.

Fixes #70.